% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewMLEboot.R
\name{estBCa}
\alias{estBCa}
\title{Estimate the adjusted bootstrap percentile (BCa)}
\usage{
estBCa(boot.est, orig_stat, CI = 0.9)
}
\description{
Calculating the 90\% adjusted bootstrap percentile (BCa). If the estimator
is sensitive to the data during bootstrapping, this fixes the skewness
and the apparent outliers of a distribution of bootstrapped estimates
}
