% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRI.R
\name{estRI}
\alias{estRI}
\title{Estimating the RLs}
\usage{
estRI(X, RI = 0.95, nBoot = NULL, ...)
}
\description{
This is a wrapper function, that estimates RIs either directly from the
data or from a previously established re-weighted RI.
}
