% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewMLE.R
\name{mseTrim}
\alias{mseTrim}
\title{Mean Square Error (MSE) of a trimmed distribution}
\usage{
mseTrim(data, trim = 0.95)
}
\description{
This function assesses how closely a trimmed version
of the input data follows a normal distribution by
comparing the observed values against expected values
derived from a normal distribution fitted to the data.
It returns the R^{2} value as a measure of this fit.
}
