% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewMLEboot.R
\name{plot.rewMLE.boot}
\alias{plot.rewMLE.boot}
\title{plot method for rewMLE.boot.ci}
\usage{
\method{plot}{rewMLE.boot}(x, CI = 0.9, trim = 0.95, ..., main = NULL)
}
\description{
plot method for rewMLE.boot.ci
}
