% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewMLE.R
\name{reweightBCr}
\alias{reweightBCr}
\title{function for reweighted maximum likelihood, based on an initial estimate.
args:
x              : vector of sorted original observations
zt.raw         : vector of sorted poststandardized transformed data
(from the initial lambdahat)
lambdahat.raw  : initial estimate for transformation parameter lambda
lambdarange    : range of lambda values.
quant          : quantile for determining the weights in the
reweighting step
nbsteps        : number of reweighting steps}
\usage{
reweightBCr(
  x,
  zt.raw,
  lambdahat.raw,
  lambdarange,
  winput = NULL,
  quant = 0.99,
  nbsteps = 2,
  standardize = TRUE
)
}
\description{
function for reweighted maximum likelihood, based on an initial estimate.
args:
x              : vector of sorted original observations
zt.raw         : vector of sorted poststandardized transformed data
(from the initial lambdahat)
lambdahat.raw  : initial estimate for transformation parameter lambda
lambdarange    : range of lambda values.
quant          : quantile for determining the weights in the
reweighting step
nbsteps        : number of reweighting steps
}
