% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewMLE.R
\name{RewML_BC}
\alias{RewML_BC}
\title{The reweighted ML estimator for the Box-Cox transformation parameter.}
\usage{
RewML_BC(
  x,
  lambdarange = NULL,
  standardize = TRUE,
  init = "BCr",
  winput = NULL,
  quant = 0.99,
  nbsteps = 2
)
}
\description{
args:
x:              vector of \emph{sorted} observations
lambdarange:    grid of lambda values. If NULL, a grid between
-2 and 4 is chosen
init:           initial estimator. should be "BCr" or "BC"
quant:         quantile for determining the weights in the
reweighting step
nbsteps:        number of reweighting steps
}
