% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewMLE.R
\name{RewML_YJ}
\alias{RewML_YJ}
\title{The reweighted ML estimator for the Yeo-Johnson transformation parameter.}
\usage{
RewML_YJ(
  x,
  lambdarange = NULL,
  quant = 0.99,
  init = "YJr",
  winput = NULL,
  standardize = TRUE,
  nbsteps = 2
)
}
\description{
args:
x             : vector of \emph{sorted} observations
lambdarange       : grid of lambda values. If NULL, a grid between
-2 and 4 is chosen
init          : initial estimator. should be "YJr" or "YJ"
quant         : quantile for determining the weights in the
reweighting step
nbsteps       : number of reweighting steps
}
