% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewMLEboot.R
\name{rewMLE.boot.ci}
\alias{rewMLE.boot.ci}
\title{Confidence Intervals for the rewMLE Bootstrapping}
\usage{
rewMLE.boot.ci(rewMLE.boot.out, CI = 0.9)
}
\description{
Confidence Intervals for the rewMLE Bootstrapping
}
