% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewMLEboot.R
\name{rewMLE.boot}
\alias{rewMLE.boot}
\title{re-weighted Maximum Likelihood Estimator (MLE) with Bootstrapping}
\usage{
rewMLE.boot(
  X,
  type = "YJ",
  standardize = TRUE,
  quant = 0.995,
  nbsteps = 2,
  winput = NULL,
  nBoot = 200
)
}
\description{
Functions rewMLE, rewMLE.boot, and rewMLE.boot.ci as custom are custom
wrappers derived from the "transfo" function of the cellWise package to
specifically adapt the re-weighted MLE for use with clinical routine data
from laboratory medicine. They contain modifications on the published
code from the cellWise package
}
\details{
Original Package Reference:
cellWise: Analyzing Data with Cellwise Outliers
Version: 2.5.2
License: GPL-2 | GPL-3 \link{expanded from: GPL (≥ 2)}
https://cran.r-project.org/web/packages/cellWise/index.html
}
