% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewRI.R
\name{rewRI}
\alias{rewRI}
\title{Estimating the RIs from a previously established re-weighted MLE object}
\usage{
rewRI(rewMLE.out, RI = 0.95)
}
\description{
This function takes as an input either an object of class
"rewMLE" or "rewMLE.boot". For estimation of RIs diretly from
clinical data, refer to the \code{estRI} function, which is extended
wrapper function of this function.
}
