% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRI.R
\name{tnormRI}
\alias{tnormRI}
\title{Reference Interval from a Normal Distribution}
\usage{
tnormRI(n = 1000, mean = 0, sd = 1, RI = 0.95, CI = 0.9)
}
\description{
Calculate the Reference Interval (+ 90\% Confidence Interval) from a normal distribution
given the population mean (mu) and standard deviation (sd, sigma)
}
\details{
Confidence intervals of sample error
se = sqrt(((sd^2)/n) + (((z.score.RI^2)\emph{(sd^2))/(2}n)))

https://en.wikipedia.org/wiki/Reference_range#Confidence_interval_of_limit
}
