% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewMLE.R
\name{transfo_rewMLE}
\alias{transfo_rewMLE}
\title{Transform data by re-weighted MLE of BC/YJ}
\usage{
transfo_rewMLE(Xnew, rewMLE.out)
}
\description{
This function takes new data (Xnew) and transform them
according to the parameters in the previously established
re-weighted MLE (rewMLE.out).
The rewMLE.out needs to an objects of class
\code{rewMLE}, \code{rewMLE.boot}, or \code{rewMLE.boot.ci}.
Source: This is a modification on the \code{transfo_newdata} function.
cellWise: Analyzing Data with Cellwise Outliers
Version: 2.5.2
https://cran.r-project.org/web/packages/cellWise/index.html
}
