% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewMLE.R
\name{transfoback_rewMLE}
\alias{transfoback_rewMLE}
\title{Back transform data by re-weighted MLE of BC/YJ}
\usage{
transfoback_rewMLE(Ynew, rewMLE.out)
}
\description{
This function back-transforms data (Ynew) and transform them
back to the original space. This is done in accordance to the parameters
in the previously established re-weighted MLE (rewMLE.out).
}
\details{
The rewMLE.out needs to an objects of class
\code{rewMLE}, \code{rewMLE.boot}, or \code{rewMLE.boot.ci}.
Source: This is a modification on the \code{transfo_transformback} function.
cellWise: Analyzing Data with Cellwise Outliers
Version: 2.5.2
https://cran.r-project.org/web/packages/cellWise/index.html
}
